+++ app/src/main/java/house/edc/pocket/EdcLockWidgetReceiver.kt
@@ -1,0 +1,8 @@
+package house.edc.pocket
+
+import androidx.glance.appwidget.GlanceAppWidget
+import androidx.glance.appwidget.GlanceAppWidgetReceiver
+
+class EdcLockWidgetReceiver : GlanceAppWidgetReceiver() {
+    override val glanceAppWidget: GlanceAppWidget = EdcLockWidget()
+}